change to float instead of int

+++ tire/carrigan_tire.py
@@ -2,7 +2,7 @@
 from tire.tire import Tire
 
 class CarriganTire(Tire):
-    def __init__(self, wearings: list[int]):
+    def __init__(self, wearings: list[float]):
         self.wearings = wearings
         self.__is_need_service = len([w for w in self.wearings if w >= 0.9]) > 1
