+++ tire/octoprime_tire.py
@@ -2,7 +2,7 @@
 from tire.tire import Tire
 
 class OctoprimeTire(Tire):
-    def __init__(self, wearings: list[int]):
+    def __init__(self, wearings: list[float]):
         self.wearings = wearings
         self.__is_need_service = sum(wearings) >= 3
